fix expander vars 0 78839
